Separate logic into separate files
a

Update of_json.rs
